viafree.fi
I made filters for videos of this site. A nice load of filters was needed for that site.... :D

Here is a sample url:
https://www.viafree.fi/ohjelmat/dokumentit/sotarikokset/kausi-1/jakso-6

However there are small "hiccups". Videos load without ads but there is a small quick flashing of "video cannot be loaded" text that however goes away quickly. That text flashes every time the sites tries to load ads (the beginning of the video and commercial breaks). But this will work anyway.