Remove /utag/sanoma-fi/ because of various issues it causes
Previously this filter caused issues on vauva.fi and now it is causing issues on https://www.vaalikone.fi/eduskunta2019/ehdokkaat/ when trying to view a candidate. It will just endlessly load. It's very likely that this filter breaks other sites also that we are not aware of so it's better to remove this filter completely.