Suggestion to remove Yle analytics
There is two known occasions #60 #61 when blocking Yle analytics has caused breaking issues on websites. This rule doesn't seem to be safe and reliable and it's very possible that it breaks other sites also we are not aware of.

And since it's not an adblocking rule either I think it's just an unnecessary risk to keep it.